bank/requirements.txt
streamlit
mysql-connector-python
reportlab
matplotlib
python-dotenv
